+++ scaffolding-way/src/components/assignments/AssignmentCreate.vue
@@ -1,14 +1,13 @@
 <script setup>
-import { ref } from 'vue'
 
 let emit = defineEmits(['newItemAdded'])
-let item = ref('')
+let item = ""
 </script>
 <template>
   <form action="#" method="post" @submit.prevent="emit('newItemAdded', item)">
     <label for="item">
-      <input id="item" v-model="item" name="item" type="text" />
-      <input type="submit" />
+      <input id="item" v-model="item" name="item" type="text"/>
+      <input type="submit"/>
     </label>
   </form>
 </template>