+++ scaffolding-way/src/views/AssignmentView.vue
@@ -1,8 +1,8 @@
 <script setup>
-import Assignments from '@/components/assignments/Assignments.vue'
-</script>
+import Assignments from '@/components/assignments/Assignments.vue'</script>
 <template>
   <main>
+    <h2 style="color: hsla(160, 100%, 37%, 1);">Assigment List</h2>
     <Assignments></Assignments>
   </main>
 </template>