+++ scaffolding-way/src/views/UserView.vue
@@ -1,0 +1,12 @@
+<script setup>
+import UserCard from "@/components/user/UserCard.vue";</script>
+
+<template>
+  <main>
+    <UserCard></UserCard>
+  </main>
+
+</template>
+
+<style scoped>
+</style>